Add Transaction middleware
