+++ src/Web3ProviderBackend.ts
@@ -94,21 +94,6 @@ export class Web3ProviderBackend extends EventEmitter implements IWeb3Provider {
 
   async _request(req: JsonRpcRequest): Promise<any> {
     switch (req.method) {
-      case 'eth_sendTransaction': {
-        const wallet = this.#getCurrentWallet()
-        const rpc = this.getRpc()
-        // @ts-expect-error todo: parse params
-        const jsonRpcTx = req.params[0]
-
-        const txRequest = convertJsonRpcTxToEthersTxRequest(jsonRpcTx)
-        try {
-          const tx = await wallet.connect(rpc).sendTransaction(txRequest)
-          return tx.hash
-        } catch (err) {
-          throw err
-        }
-      }
-
       // todo: use the Wallet Permissions System (WPS) to handle method
       case 'wallet_requestPermissions': {
         if (
@@ -289,51 +274,4 @@ function without<T>(list: T[], item: T): T[] {
     return list.slice(0, idx).concat(list.slice(idx + 1))
   }
   return list
-}
-
-// Allowed keys for a JSON-RPC transaction as defined in:
-// https://ethereum.github.io/execution-apis/api-documentation/
-const allowedTransactionKeys = [
-  'accessList',
-  'chainId',
-  'data',
-  'from',
-  'gas',
-  'gasPrice',
-  'maxFeePerGas',
-  'maxPriorityFeePerGas',
-  'nonce',
-  'to',
-  'type',
-  'value',
-]
-
-// Convert a JSON-RPC transaction to an ethers.js transaction.
-// The reverse of this function can be found in the ethers.js library:
-// https://github.com/ethers-io/ethers.js/blob/v5.7.2/packages/providers/src.ts/json-rpc-provider.ts#L701
-function convertJsonRpcTxToEthersTxRequest(tx: {
-  [key: string]: any
-}): ethers.providers.TransactionRequest {
-  const result: any = {}
-
-  allowedTransactionKeys.forEach((key) => {
-    if (tx[key] == null) {
-      return
-    }
-
-    switch (key) {
-      // gasLimit is referred to as "gas" in JSON-RPC
-      case 'gas':
-        result['gasLimit'] = tx[key]
-        return
-      // ethers.js expects `chainId` and `type` to be a number
-      case 'chainId':
-      case 'type':
-        result[key] = Number(tx[key])
-        return
-      default:
-        result[key] = tx[key]
-    }
-  })
-  return result
 }
